refactor: split frontend types by domain
